Update code generation for unsupported algorithms
The generate.py script was updated for the introduced 'kex-names.c'
and 'sshd-session.c` files. OQS template subdirectories and requesite
templates were added for these files as well. There was also a change to
the README fragment to prevent adding a '-' character to the MLKEM algs,
avoiding confusion to the old codepoint references being added back when
the generate script is run.

Signed-off-by: Andrew Younkers <ayounkers44@gmail.com>